fix bot create
